Split prod/stage configs, remove testnets
