+++ network.config/index.ts
@@ -1,0 +1,5 @@
+export * from "./types";
+export * from "./ids";
+export {prodNetworkConfig} from "./prod";
+export {stageNetworkConfig} from "./stage";
+export {repayerConfig} from "./repayer";
