+++ specific-fork-test/unichain/RepayerUSDT0.ts
@@ -14,7 +14,7 @@ import {
   TransparentUpgradeableProxy, ProxyAdmin,
   TestLiquidityPool, Repayer,
 } from "../../typechain-types";
-import {networkConfig} from "../../network.config";
+import {prodNetworkConfig as networkConfig} from "../../network.config";
 
 describe("Repayer USDT0 (Unichain fork)", function () {
   const deployAll = async () => {
